New public method in OutOfCoreOptions.
The new method adds a new hint in the *Database Registry* cache, to let caching of more objects. By default, the *Registry* only caches **ARCHIVES** node types.

I had encounter an issue with the PagedLOD and the point clouds. For example, imagine a point cloud from a town road. If I set an AnimationPath that simulates that road and move the cloud alongside, with a high zoom level (i.e.: close to the cloud) several LODs will be loaded.

When reaching the end of the path and going back to the first spot, if the path is re-run, the LODs are still in memory but OSG won't recognize them so they will be loaded again, consuming more memory. Of course, setting the max LOD target in the Database Pager to a low number also helps.