Body for the new method
It is called in the constructors with a *CACHE_NODES* parameter that it's added to the default parameter in the Options (CACHE_ARCHIVES).

This will let the caching of the model LOD nodes. With the standard LOD target in the DatabasePager (300) I scarcely found any gain. I set it to a very low value (for example, 10) and my memory consuption stopped to increase while playing with the cloud.

This worked for me, but maybe a better/more elaborate solution will be helpful.